change server members to a dict
Considering that we frequently find members by their ID, it makes sense
to store this as a dict where the keys are the IDs so we can look them
up (and add and remove them) in O(1) time instead of O(N) time.

Keeps backwards compatibility so you can still do server.members and
get a list (it just stores it as a dict internally).