Implement rate limit header support
We will keep track of the headers in a per bucket dict. None key dict
is the default rate limits. We don't increment or decrement anything and
rely on the server sending us the informaiton needed.

We use asyncio.Event and set the flag. We clear the flag when we hit the
global rate limit and set it again when we can continue making requests.
This will allow the wait() coroutine to return immediately.

This should satisfy hammerandchisel/discord-api-docs#108