Allowing a function to execute right before a command is invoked
This adds an optional parameter `before_invoke` of the type `function` to the `process_commands` coroutine. If `before_invoke` is given as an argument to `process_commands`, it will execute that function before a command is invoked (if `message` contains a command). This is useful because listening to `on_command` doesn't guarantee that the code that should be executed `on_command` is actually executed before the command is invoked.