Embed.set_image/author/thumbnail don't set None
Before this change Embed.set_image(None) would append the string 'None'
to the request, leading to an opaque 400 Bad Request error introduced by
discord.py, not the user. This change detects None entered for the URL
in set_image and set_thumbnail and does not append that string 'None'.
For consistency, it also changes set_author to ignore None as well,
although the Discord API accepts the string 'None' as the author in that
case and does not produce a 400 Error.