Edited to work with rewrite models.
I was rebuilding my music bot and noticed much of this code was using the old models. This is my current, working copy of the code that takes requests from users.

My queue-handling and next song playing is probably hacky, so any improvements I could make there would be appreciated.