Added an enum representing the valid activity types.
I have added this, as it does not seem to be defined anywhere else
in this API wrapper. I have grepped all the files to try and find
a match, but as yet can not find anything.

Currently, the existing values for the game activity types are not
very clear. There does not seem to be anything in the dpy
documentation regarding this (at least, not in the docstrings in
the source code).

The docstring in discord.Game seems to imply that there is only
the option of playing a game, or streaming a twitch URL. As of
4th Jan 2018, this is incorrect. There is also the ability to
be "listening to" something, or "watching" something. Both have
been tested and work running under a standard discord bot account.

I have tweaked the docstring in discord.Game to reflect this change.

EDIT (--amend):
I fixed the docstring in Game. Seems ST3 didn't save properly.