Add consume_args option to commands.group
When invoke_without_command is set to False (default), the callback
function is called before the subcommand callback. If the callback has
arguments, they will be consumed from the command, and the subcommand
will be looked for in whatever is left. This adds a consume_args option,
which when True (the default), keeps this functionality. When set to
False, however, the arguments are passed to the callback, but not
consumed. This means that the callback can perform whatever setup or
validation it likes on the arguments, and then the subcommand will still
be called properly.