Default invoke_without_command to True instead of False
The current implementation is at best kind of unexpected for how one would expect subcommands to work- if the second arg of a command is a subcommand, it's natural to assume that it will invoke the subcommand instead of being passed as an arg to the root command. The default option also seems to be a more specialty option, with less total usage by the average user. 

Furthermore, the naming `invoke_without_command` is... confusing at best. I'm still not completely sure what this means. This change should eliminate some minor confusion for new users.