Fix behavior of everyone role mention
Prior to this change, attempting to do `everyone_role.mention` would result in the mention being formatted strangely when viewed in discord, appearing as `@@everyone`. 

This fixes this behavior, which appears as a kind of "gotya" to anyone trying to make a message filter ala #1429, or trying, for some reason, to list off all a members roles including everyone.