Make Guild.features a frozenset
The most frequent use-case for Guild.features is containment checks, which are slow on a list.
The API does not guarantee any order for guild features anyway, so it's unlikely that anyone would use indexing on this
list. Additionally, using a set allows for use cases such as the following, which checks if a given guild is partnered:

def is_partnered(guild):
	partnered_features = {'VIP_REGIONS', 'VANITY_URL', 'INVITE_SPLASH'}
	return guild.features.issuperset(partnered_features)

Finally, a frozenset is used instead of a set because it should not be modified.