Removed requirements for messenger
Checking for a user's humanity should be a users choice and thus it is added as a default global check. Should the user want to enable or disable this for testing purposes or for only specific bots for testing, it should be up to them.

Documentation for this feature is also updated.