Update bot.py
Changed wording for documentation so that it was easier to understand. Made check closer to implementation of that documentation.

Added note for this commit that I didn't think of last commit. This actually makes self_bot useful because the previous check would never allow for the command from the bot to be processed as the bot is, of course, a bot.