Update Member.top_role
In certain cases (presumably from a caching issue), the roles property will not be sorted correctly so the top_role function will not work correctly as it just returns the last role in the list (which should have the highest position). Changing the top_role function to search through all the roles and find the role with the highest position and return that instead fixes the function in these specific events.