Make Embed.copy() a deep copy
I'm requesting that Embed.copy become a less shallow copy. The only difference here is that the fields will be copied by value instead of the whole list reference, because that's the only thing that persists between copies.

small repro:

- make embed
- loop
	- clear fields
	- add some fields
	- yield embed.copy()

all yielded embeds would have the final embed's fields.

---

I removed `self.__slots__` and replaced it with a hard-coded `'_field'`-less tuple of __slots__. There are multiple ways to do what I'm trying to solve here and this may not have been the cleanest, so I'm all ears for suggestions.