raise InvalidArgument when using overwrites
when creating a channel with the overwrites set but using the wrong type
for the keys a not so descriptive AttributeError was raised as the
assumed to be a Member target does not have an id attribute.

instead the target is checked and a clear InvalidArgument is raised when
the overwrites dict has the wrong class as a key.