Use channel default auto archive duration
Currently creating a new thread uses a default auto archive duration of
1440 minutes, or 1 day.

Rather than prescribing our own default, we can use the default auto
archive duration that is set on the channel. This ensures that newly
created threads will respect the default auto archive duration as
prescribed by the user.

My initial approach was to simply omit the auto_archive_duration from
the request, but when doing so, the thread is always created with an
auto archive duration of 1440 minutes, regardless of the channel's
default setting.

See the following PR for more discussion on the matter:

  https://github.com/discord/discord-api-docs/pull/3567

It's unclear if this is intended behaviour or not, but we might as well
work with it for the time being