Use get_bot_mapping to generate bot help
This change makes it so that `DefaultHelpCommand` uses it's `mapping` argument (generated by `get_bot_mapping()`), to generate it's help output. This change does not directly change DefaultHelpCommand behavior and will continue to generate the same output, but allows subclasses to override get_bot_mapping to make alterations to the return value of get_bot_mapping to, for example, add or remove elements.