Fix: Piped Audio Input Ends Prematurely #9001
This commit addresses Issue #9001 by replacing
`self._process.terminate()` with `self._stdin.close()` to fix the issue
where the process was being interrupted after reading input.
I also add an `assert self._stdin is not None` statement to resolve
the type of `self._stdin` and remove the redundant if statement.